+++ R/element.R
@@ -22,10 +22,13 @@ setClass(
 #'
 #' Make a new `R2PptxElement`. Element represent text, a flextable, an image, a
 #' ggplot2, etc to add to a slide.
-#' @export
+#'
+#' @param key character. Name of the placeholder label for this element.
+#' @param value object. Object to put into a PowerPoint slide, eg text or a plot.
 #' @return An object of class \code{R2PptxElement} representing something to put
 #'   on a slide.
 setGeneric("new_element", function(key, value) standardGeneric("new_element"))
+
 
 #' New Element
 #'
@@ -44,6 +47,7 @@ setMethod(
     new("R2PptxElement", key = location, value = value)
   }
 )
+
 
 #' New Element
 #'
@@ -82,20 +86,4 @@ setMethod(
   "location",
   "R2PptxElement",
   function(x) x@key
-)
-
-
-# TODO rename this generic / method
-setGeneric("add_pptx", function(x, pptx_obj) standardGeneric("add_pptx"))
-setMethod(
-  "add_pptx",
-  "R2PptxElement",
-  function(x, pptx_obj) {
-    location <- location(x)
-    officer::ph_with(
-      pptx_obj,
-      value = value(x),
-      location = do.call(ph_location_fn(location), ph_location_args(location))
-    )
-  }
 )